refactor: move dashboard to page package and rename MetricsDao.java to MetricsAccessor.java
