+++ src/main/java/com/github/khakers/modmailviewer/page/dashboard/ChartData.java
@@ -1,4 +1,4 @@
-package com.github.khakers.modmailviewer.dashboard;
+package com.github.khakers.modmailviewer.page.dashboard;
 
 public record ChartData<T, U>(T[] data, U[] labels) {
 }
